+++ resources/payout_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/razorpay/razorpay-go/constants"
-	"github.com/razorpay/razorpay-go/resources"
 	"github.com/razorpay/razorpay-go/utils"
 	"github.com/stretchr/testify/assert"
 )
@@ -40,36 +39,6 @@ func TestPayoutFetch(t *testing.T) {
 		"expand[]": "fund_account",
 	}
 	body, err := utils.Client.Payout.Fetch(TestPayoutID, queryParams, nil)
-	assert.Nil(t, err)
-	assert.NotNil(t, body)
-
-	jsonByteArray, _ := json.Marshal(body)
-	utils.TestResponse(jsonByteArray, []byte(fixture), t)
-}
-
-// TestPayoutCreate is a function to test creating a new payout
-func TestPayoutCreate(t *testing.T) {
-	url := "/" + constants.VERSION_V1 + constants.PAYOUT_URL
-	teardown, fixture := utils.StartMockServer(url, "fake_payout")
-	defer teardown()
-
-	req := &resources.PayoutRequest{
-		AccountNumber:     "7878780080316316",
-		FundAccountID:     "fa_00000000000001",
-		Amount:            1000000,
-		Currency:          "INR",
-		Mode:              "IMPS",
-		Purpose:           "refund",
-		QueueIfLowBalance: true,
-		ReferenceID:       "Acme Transaction ID 12345",
-		Narration:         "Acme Corp Fund Transfer",
-		Notes: map[string]string{
-			"notes_key_1": "Tea, Earl Grey, Hot",
-			"notes_key_2": "Tea, Earl Grey… decaf.",
-		},
-	}
-
-	body, err := utils.Client.Payout.Create(req, nil)
 	assert.Nil(t, err)
 	assert.NotNil(t, body)