+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.3] - 2025-05-02
+feat: Added new API endpoints
+- Added Payout API:
+  - Fetch all payouts
+  - Fetch specific payout by ID
+  
 ## [1.3.2] - 2024-04-17
 feat: Added new API endpoints